requirements.txt
Django~=5.0.5
djangorestframework~=3.15.1
python-dotenv~=1.0.1
asgiref==3.7.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
django-cors-headers==4.3.1
djoser==2.2.2
Pillow==10.3.0
psycopg2-binary==2.9.9
cachetools==5.3.3